requirements.txt
google-cloud-aiplatform[agent_engines]==1.95.1
cloudpickle==3.0.0
pydantic==2.11.7
python-dotenv
requests
pathlib
protobuf==3.20.3